requirements.txt
streamlit==1.38.0
pandas==2.2.2
mysql-connector-python==8.0.33
openpyxl==3.1.2
plotly==5.22.0
sqlalchemy==2.0.30
matplotlib==3.7.1
seaborn==0.13.2
numpy==1.24.3
pyarrow==14.0.1
fastparquet==2023.10.0
scikit-learn==1.3.0
python-dotenv==1.0.0
scipy==1.11.0